requirements.txt
praw==7.7.1
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
nltk==3.8.1
python-dotenv==1.0.0
flask==2.3.3
werkzeug==2.3.7
matplotlib==3.7.1
seaborn==0.12.2
textblob==0.17.1 